fix: rename tailwind related files
