+++ packages/jokul/src/core/index.ts
@@ -8,4 +8,4 @@ export type {
     WithOptionalChildren,
 } from "./types.js";
 export { default as tokens } from "./tokens.js";
-export { default as tailwindPresets } from "./tailwind/tailwindConfig.js";
+export { default as tailwindPreset } from "./tailwind/tailwindPreset.js";
